move working_bank_entry and error to separate module
