+++ banking-bench/src/main.rs
@@ -21,7 +21,10 @@ use {
     },
     solana_measure::measure::Measure,
     solana_perf::packet::{to_packet_batches, PacketBatch},
-    solana_poh::poh_recorder::{create_test_recorder, PohRecorder, WorkingBankEntry},
+    solana_poh::{
+        poh_recorder::{create_test_recorder, PohRecorder},
+        working_bank_entry::WorkingBankEntry,
+    },
     solana_runtime::{
         bank::Bank, bank_forks::BankForks, prioritization_fee_cache::PrioritizationFeeCache,
     },
