+++ core/src/banking_stage/consumer.rs
@@ -16,9 +16,11 @@ use {
     solana_fee::FeeFeatures,
     solana_ledger::token_balances::collect_token_balances,
     solana_measure::{measure::Measure, measure_us},
-    solana_poh::poh_recorder::{
-        BankStart, PohRecordError, RecordTransactionsSummary, RecordTransactionsTimings,
-        TransactionRecorder,
+    solana_poh::{
+        poh_record_error::PohRecordError,
+        poh_recorder::{
+            BankStart, RecordTransactionsSummary, RecordTransactionsTimings, TransactionRecorder,
+        },
     },
     solana_runtime::{
         bank::{Bank, LoadAndExecuteTransactionsOutput},
