+++ poh/src/lib.rs
@@ -1,7 +1,9 @@
 #![allow(clippy::arithmetic_side_effects)]
 pub mod leader_bank_notifier;
+pub mod poh_record_error;
 pub mod poh_recorder;
 pub mod poh_service;
+pub mod working_bank_entry;
 
 #[macro_use]
 extern crate solana_metrics;
