+++ poh/src/working_bank_entry.rs
@@ -1,0 +1,3 @@
+use {solana_entry::entry::Entry, solana_runtime::bank::Bank, std::sync::Arc};
+
+pub type WorkingBankEntry = (Arc<Bank>, (Entry, u64));
